Base

+++ Assets/01.Scripts/Combat/IDamageable.cs
@@ -5,6 +5,6 @@ namespace Combat
 
     interface IDamageable
     {
-        public void ApplyDamage(int damage);
+        public void ApplyDamage(float damage);
     }
 }
